--- types/isolated-web-apps/dprint.json
@@ -1,15 +1,0 @@
-{
-  "excludes": [
-    "**/node_modules",
-    "**/*-lock.json"
-  ],
-  "plugins": [
-    "https://plugins.dprint.dev/typescript-0.95.12.wasm",
-    "https://plugins.dprint.dev/json-0.21.0.wasm",
-    "https://plugins.dprint.dev/markdown-0.20.0.wasm",
-    "https://plugins.dprint.dev/toml-0.7.0.wasm",
-    "https://plugins.dprint.dev/g-plane/malva-v0.15.0.wasm",
-    "https://plugins.dprint.dev/g-plane/markup_fmt-v0.24.1.wasm",
-    "https://plugins.dprint.dev/g-plane/pretty_yaml-v0.5.1.wasm"
-  ]
-}
