--- types/isolated-web-apps/package.json
@@ -1,20 +1,0 @@
-{
-  "private": true,
-  "nonNpm": true,
-  "nonNpmDescription": "Implementation of Isolated Web App (IWA) APIs resides in Chromium.",
-  "name": "@types/isolated-web-apps",
-  "version": "0.0.9999",
-  "projects": [
-    "https://github.com/WICG/isolated-web-apps"
-  ],
-  "devDependencies": {
-    "@types/isolated-web-apps": "workspace:.",
-    "@types/web": "*"
-  },
-  "owners": [
-    {
-      "name": "Paulina Gacek",
-      "githubUsername": "paulinagacek"
-    }
-  ]
-}
