--- types/isolated-web-apps/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "node16",
-    "lib": [
-      "es2020",
-      "dom"
-    ],
-    "noImplicitAny": true,
-    "noImplicitThis": true,
-    "strictFunctionTypes": true,
-    "strictNullChecks": true,
-    "types": [],
-    "noEmit": true,
-    "forceConsistentCasingInFileNames": true
-  },
-  "files": [
-    "index.d.ts",
-    "screencapture-tests.ts",
-    "direct-sockets-tests.ts",
-    "smart-card-tests.ts",
-    "webusb-tests.ts",
-    "controlled-frame-tests.ts"
-  ]
-}
