+++ types/isolated-web-apps__controlled-frame/isolated-web-apps__controlled-frame-tests.ts
@@ -1,6 +1,7 @@
 import {
   ClearDataOptions,
   ClearDataTypeSet,
+  ConsoleMessage,
   ConsoleMessageEvent,
   ContentScriptDetails,
   ContextMenus,
@@ -8,22 +9,29 @@ import {
   ContextMenusCreateProperties,
   DialogController,
   DialogEvent,
+  DialogMessage,
   DialogType,
   HTMLControlledFrameElement,
   ImageDetails,
   InjectDetails,
   LoadAbortEvent,
+  LoadAbortInfo,
   LoadRedirectEvent,
+  LoadRedirectInfo,
+  NewWindow,
   NewWindowController,
   NewWindowEvent,
+  PermissionRequest,
   PermissionRequestControllerBase,
   PermissionRequestEvent,
   PermissionType,
   RequestedHeaders,
   ResourceType,
   RunAt,
+  SizeChange,
   SizeChangedEvent,
   WebRequest,
+  URLPattern,
   WebRequestAuthCredentials,
   WebRequestAuthOptions,
   WebRequestAuthRequiredEvent,
@@ -33,9 +41,10 @@ import {
   WebRequestInterceptorOptions,
   WebRequestResponse,
   WindowOpenDisposition,
+  ZoomChange,
   ZoomChangeEvent,
   ZoomMode,
-} from "isolated-web-apps";
+} from "isolated-web-apps/controlled-frame";
 
 const dummyControlledFrame: HTMLControlledFrameElement = {} as HTMLControlledFrameElement;
 const dummyDialog: DialogController = {} as DialogController;
