+++ types/isolated-web-apps__controlled-frame/tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "compilerOptions": {
+        "module": "node16",
+        "lib": [
+            "es6",
+            "dom"
+        ],
+        "noImplicitAny": true,
+        "noImplicitThis": true,
+        "strictFunctionTypes": true,
+        "strictNullChecks": true,
+        "types": [],
+        "noEmit": true,
+        "forceConsistentCasingInFileNames": true,
+        "paths": {
+            "isolated-web-apps/controlled-frame": [
+                "./index.d.ts"
+            ]
+        }
+    },
+    "files": [
+        "index.d.ts",
+        "isolated-web-apps__controlled-frame-tests.ts"
+    ]
+}
