+++ types/isolated-web-apps__direct-sockets/.npmignore
@@ -1,0 +1,5 @@
+*
+!**/*.d.ts
+!**/*.d.cts
+!**/*.d.mts
+!**/*.d.*.ts
