+++ types/isolated-web-apps__smart-card/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+    "rules": {
+        "@typescript-eslint/no-empty-interface": "off"
+    }
+}
