+++ types/isolated-web-apps__smart-card/tsconfig.json
@@ -1,0 +1,25 @@
+{
+    "compilerOptions": {
+        "module": "node16",
+        "lib": [
+            "es6",
+            "dom"
+        ],
+        "noImplicitAny": true,
+        "noImplicitThis": true,
+        "strictFunctionTypes": true,
+        "strictNullChecks": true,
+        "types": [],
+        "noEmit": true,
+        "forceConsistentCasingInFileNames": true,
+        "paths": {
+            "isolated-web-apps/smart-card": [
+                "./index.d.ts"
+            ]
+        }
+    },
+    "files": [
+        "index.d.ts",
+        "isolated-web-apps__smart-card-tests.ts"
+    ]
+}
